requirements.txt
python-dotenv==0.21.0
requests==2.28.1
groq==0.1.0          
pytest==7.1.2
